requirements.txt
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
typing-extensions>=4.5.0
requests>=2.31.0
ta>=0.10.2
alpaca-trade-api>=3.1.1
vectorbt>=0.25.2